requirements.txt
tk
ttkbootstrap
pillow
pycryptodome
stepic
zlib-ng
stegano